generate random room name if not already specified
Previously, any URI (including the root path) was considered a valid
room name, which meant anyone going there by default would see anyone
else who happened to be there.  That can make for fun surprises, but
it's probably not a great default.  Instead, we now generate a new,
random room name automatically and redirect to it.  The user may then
copy the URL and send it to invitees.

Also, we now use the entire URL as the full room name so that separate
servers are isolated from each other even if they share the same
backend database.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>